pyproject.toml
[project]
name = "ai-worthy-api-roo-1"
version = "0.1.0"
description = "A financial tracking API built with FastAPI and SQLite"
authors = [
    {name = "readyyyk",email = "mr.sairexer@gmail.com"}
]
readme = "README.md"
requires-python = "3.13.2"
dependencies = [
    "fastapi>=0.109.0",
    "sqlalchemy>=2.0.25",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "uvicorn>=0.25.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.6",
    "greenlet (>=3.1.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "ai_worthy_api_roo_1", from = "src"}]

[tool.poetry.group.test.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.1"
httpx = ">=0.24.1"
pytest-cov = ">=4.1.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
